test(checkpoint): improve snapshot diff reporting and fix recipe seed modulo (LIM-1223)
- Avoid 32-bit truncation before modulo in Recipe::from_seed
- Detail differing fields and full snapshot context in checkpoint mismatch error messages
- Fold tick check into snapshot comparison

+++ tests/checkpoint_resume/harness.rs
@@ -51,7 +51,7 @@ impl Recipe {
     ];
 
     pub(crate) fn from_seed(seed: u64) -> Self {
-        Self::ALL[(seed as usize) % Self::ALL.len()]
+        Self::ALL[(seed % Self::ALL.len() as u64) as usize]
     }
 }
